Add quiz controller

+++ quiz-service/src/main/java/net/javaguides/quizservice/repository/QuizRepository.java
@@ -3,5 +3,5 @@ package net.javaguides.quizservice.repository;
 import net.javaguides.quizservice.model.Quiz;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface QuizRepository extends JpaRepository<Quiz, Long> {
+public interface QuizRepository extends JpaRepository<Quiz, Integer> {
 }
